refactor(profile): route; component
UPDATED:
- profile: routing module
- profile-filter: component
- profile-form-modal: component
- profile-list: component

+++ src/app/features/profile/profile-list/profile-list.component.ts
@@ -1,10 +1,11 @@
-import { NotificationService } from './../../../services/notification.service';
 import { ProfileLocalService } from '../profile.local.service';
 import { Component, Input, OnInit } from '@angular/core';
 import { IProfile, Profile } from 'src/app/interfaces/Profile';
 import { EventEmitterService } from 'src/app/services/event-emitter.service';
 import { ProfileService } from 'src/app/services/profile.service';
 import { UtilService } from 'src/app/services/util.service';
+import { BsModalRef, BsModalService } from 'ngx-bootstrap/modal';
+import { ProfileFormModalComponent } from '../profile-form-modal/profile-form-modal.component';
 
 @Component({
     selector: 'app-profile-list',
@@ -14,35 +15,62 @@ import { UtilService } from 'src/app/services/util.service';
 export class ProfileListComponent implements OnInit {
 
     @Input() profileList: IProfile[];
+    @Input() filter: IProfile;
+
+    public contextFilter: string;
+
+    private bsModalRef: BsModalRef;
 
     constructor(
         public profileLocalService: ProfileLocalService,
         public profileService: ProfileService,
-        private notify: NotificationService
+        private bsModalService: BsModalService
     ) { }
 
     ngOnInit(): void {
-        this.index();
+        if (!this.filter) this.filter = {};
+
+        this.index(this.filter);
 
         EventEmitterService.get('is-success').subscribe((isSuccess) => {
-            if (isSuccess)
-                this.index();
+            if (!isSuccess) return;
+            this.index(this.filter);
+        });
+
+        EventEmitterService.get('profile-context-filter').subscribe((context) => {
+            this.contextFilter = context
+            console.log("context", context);
         });
     }
 
-    index(): void {
-        this.profileService.read().subscribe((data) => {
-            this.profileList = data
-            console.log("data", data.length);
-        })
+    index(filter): void {
+        this.profileService.readFilter(filter).subscribe((data) => {
+            this.profileList = data;
+        });
     }
 
     edit(profile: IProfile | null) {
         ProfileLocalService.profile = (profile) ? profile : new Profile();
     }
 
-    delete(id): void {
+    openProfileFormModal(profile: IProfile) {
+        const initialState = { profile };
+        this.bsModalRef = this.bsModalService.show(ProfileFormModalComponent, { id: UtilService.getRandom9Digits(), class: 'modal-lg', initialState });
+        this.bsModalRef.content.closeBtnName = 'Close';
+    }
+
+    delete(id: string): void {
+        if (!UtilService.isConfirm()) return;
+
         UtilService.default(this.profileService.delete(id));
+    }
+
+    changeStatus(profile: IProfile): void {
+        if (!UtilService.isConfirm()) return;
+
+        const p: IProfile = { _id: profile._id, status: !profile.status }
+
+        UtilService.default(this.profileService.update(p));
     }
 
 }
